requirements.txt
python-telegram-bot
nsetools
bsedata
yahoo_fin
yfinance
matplotlib